Cast certain playable information int values to long to prevent class cast exceptions

+++ core/src/main/java/de/danoeh/antennapod/core/service/playback/PlaybackService.java
@@ -1336,7 +1336,7 @@ public class PlaybackService extends MediaBrowserServiceCompat {
 
         if (info.playable != null) {
             Intent i = new Intent(whatChanged);
-            i.putExtra("id", 1);
+            i.putExtra("id", 1L);
             i.putExtra("artist", "");
             i.putExtra("album", info.playable.getFeedTitle());
             i.putExtra("track", info.playable.getEpisodeTitle());
@@ -1345,8 +1345,8 @@ public class PlaybackService extends MediaBrowserServiceCompat {
             if (queue != null) {
                 i.putExtra("ListSize", queue.size());
             }
-            i.putExtra("duration", info.playable.getDuration());
-            i.putExtra("position", info.playable.getPosition());
+            i.putExtra("duration", (long) info.playable.getDuration());
+            i.putExtra("position", (long) info.playable.getPosition());
             sendBroadcast(i);
         }
     }
